apps/camera-service/requirements.txt
opencv-python>=4.8.0
paho-mqtt>=1.6.0
orjson>=3.9.0
Pillow>=10.0.0
Flask>=2.3.0